Move annotations types into the `annotations` package
