+++ client/src/main/java/org/spine3/base/Identifiers.java
@@ -24,7 +24,7 @@ import com.google.common.reflect.TypeToken;
 import com.google.protobuf.Any;
 import com.google.protobuf.Message;
 import com.google.protobuf.MessageOrBuilder;
-import org.spine3.Internal;
+import org.spine3.annotations.Internal;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
