+++ client/src/main/java/org/spine3/type/TypeUrl.java
@@ -29,8 +29,8 @@ import com.google.protobuf.Descriptors.EnumDescriptor;
 import com.google.protobuf.Descriptors.GenericDescriptor;
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.google.protobuf.Message;
-import org.spine3.Internal;
 import org.spine3.annotations.AnnotationsProto;
+import org.spine3.annotations.Internal;
 import org.spine3.base.Command;
 import org.spine3.base.Event;
 import org.spine3.envelope.CommandEnvelope;
