+++ server/src/main/java/org/spine3/server/aggregate/AggregateStorage.java
@@ -22,7 +22,7 @@ package org.spine3.server.aggregate;
 
 import com.google.common.base.Optional;
 import com.google.protobuf.Timestamp;
-import org.spine3.SPI;
+import org.spine3.annotations.SPI;
 import org.spine3.base.Event;
 import org.spine3.base.EventContext;
 import org.spine3.server.storage.AbstractStorage;
