+++ server/src/main/java/org/spine3/server/entity/idfunc/GetTargetIdFromCommand.java
@@ -22,7 +22,7 @@ package org.spine3.server.entity.idfunc;
 
 import com.google.common.base.Optional;
 import com.google.protobuf.Message;
-import org.spine3.Internal;
+import org.spine3.annotations.Internal;
 import org.spine3.base.CommandContext;
 import org.spine3.server.error.MissingEntityIdException;
