+++ server/src/main/java/org/spine3/server/entity/status/package-info.java
@@ -26,6 +26,6 @@
 @ParametersAreNonnullByDefault
 package org.spine3.server.entity.status;
 
-import org.spine3.SPI;
+import org.spine3.annotations.SPI;
 
 import javax.annotation.ParametersAreNonnullByDefault;
