+++ server/src/main/java/org/spine3/server/event/EventStorage.java
@@ -20,7 +20,7 @@
 
 package org.spine3.server.event;
 
-import org.spine3.SPI;
+import org.spine3.annotations.SPI;
 import org.spine3.base.Event;
 import org.spine3.base.EventId;
 import org.spine3.server.storage.AbstractStorage;
