+++ server/src/main/java/org/spine3/server/event/storage/package-info.java
@@ -26,6 +26,6 @@
 @ParametersAreNonnullByDefault
 package org.spine3.server.event.storage;
 
-import org.spine3.SPI;
+import org.spine3.annotations.SPI;
 
 import javax.annotation.ParametersAreNonnullByDefault;
