+++ server/src/main/java/org/spine3/server/projection/ProjectionStorage.java
@@ -22,7 +22,7 @@ package org.spine3.server.projection;
 
 import com.google.common.base.Optional;
 import com.google.protobuf.Timestamp;
-import org.spine3.SPI;
+import org.spine3.annotations.SPI;
 import org.spine3.server.entity.EntityRecord;
 import org.spine3.server.storage.RecordStorage;
