+++ server/src/main/java/org/spine3/server/stand/StandStorage.java
@@ -21,7 +21,7 @@ package org.spine3.server.stand;
 
 import com.google.common.collect.ImmutableCollection;
 import com.google.protobuf.FieldMask;
-import org.spine3.SPI;
+import org.spine3.annotations.SPI;
 import org.spine3.server.entity.EntityRecord;
 import org.spine3.server.storage.RecordStorage;
 import org.spine3.type.TypeUrl;
