+++ server/src/main/java/org/spine3/util/Environment.java
@@ -23,7 +23,7 @@ package org.spine3.util;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Optional;
 import com.google.common.base.Throwables;
-import org.spine3.SPI;
+import org.spine3.annotations.SPI;
 
 import javax.annotation.Nullable;
